pyproject.toml
[tool.poetry]
name = "trantor-ai-assignment"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
sqlmodel = "^0.0.14"
aiosqlite = "^0.17.0"
backoff = "^2.2.1"
celery = "^5.3.0"
redis = "^3.5.3"
kombu = "^5.2.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
openai = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[tool.poetry.scripts]
test = "pytest"
